# Improve conversion performance for edge case ebooks