I've been using Calibre and its ebook conversion feature for a while now and everything worked fine for me until I came upon some epub file (of only 80 pages) that seems to freeze the conversion process. I ended up running the conversion during the night with the command line tool:
```
$ time ebook-convert weird.epub weird.mobi
[... calibre output ...]

real	269m59,303s
user	269m18,286s
sys	0m3,682s
```

After a bit of searching on what Calibre was doing when converting the file, it was spending most of its time creating `Stylizer` objects and, more precisely, generating rules from stylesheets. After taking a look at the html content of the ebook, it seems each word has its own inline style... ending up in generating around a thousand rules each time a stylizer object is created.

I worked on a change to move the rules creation process in a separate class and store the created object in the profile. This object gets reused by the `Stylizer` class if the stylesheets are the same. After the change, the conversion time is way better:
```
$ time ebook-convert weird.epub weird.mobi
[... calibre output ...]

real	9m10,231s
user	9m8,996s
sys	0m0,988s
```

The generated content looks the same (though I'm not sure how I could test if it's the exact same...). I also made some tests with "normal" ebooks and it seems to speed things up a bit:

Before the change:
```
$ time ebook-convert normal.epub normal.mobi
[... calibre output ...]

real	0m29,285s
user	0m29,106s
sys	0m0,176s
```
After the change:
```
$ time ebook-convert normal.epub normal.mobi 
[... calibre output ...]

real	0m24,181s
user	0m24,008s
sys	0m0,160s
```
I also tried running conversions from the GUI and it works fine too.

One possible drawback with this change is the rules are stored in the profiles and those are not destroyed after the conversion process... it may be a problem if the user converts from and to many different profiles in the same session, as the last generated rules for all those profiles will stay in memory. That may be fixed by removing the attributes from the profiles at the end of the conversion process (in `Plumber`).

But before going any further with this change, I wanted to make sure you'd be interested in this kind of changes, as it's kind of an edge case :)