@kovidgoyal 2019-05-05
I am definitely interested in patches to improve performance. Sadly I am traveling for the next three weeks so I may not have the time to review this code in the near future. Off the top of my head, why store them in the profile objects? why not store them in the oeb object? This represents the book and so has the same lifetime as the book, which is what we want.

From a quick look at the code, one thing that is wrong is you cannot compare stylesheets as sets, because their order matters. This affects CSS precedence and cascading rules. Only if the order is the same will flattened "rules" be the same.