@princesse-framboise 2019-05-05
Thanks for the quick feedback :)

I updated my code following your advices:

- the rules are now stored in the oeb object, it does make much more sense, indeed;
- I added a comparison function to test wether rules should be generated again, it test:
    - the opts object (will always be the same, as I guess it cannot change during a conversion... this is a test on object identity... it might be better to test if properties changed, but again... it's not supposed to change at all),
    - the profile object (this will change, from "input" to "output" during the process, again it's a test on object identity, but I guess profile ojects are immutable anyway)
    - and the stylesheets (and this time, I'm looping over all individual stylesheets).