@eli-schwartz 2019-05-05
> The generated content looks the same (though I'm not sure how I could test if it's the exact same...).

From a structural perspective rather than a visual perspective, you could use `calibre-debug --diff`. This will show you how the HTML markup different between the two files (convert to an epub to test).

It would be interesting to see how this affects the markup, especially if it results in fewer rules being generated as that is also less work for e-reader software. If the markup is identical then this change would be very easy to justify as a pure performance speedup.