@princesse-framboise 2019-05-06
I did some tests with `calibre-debug` and found a major flaw in my code: stylizer objects were using references to the global "rules" object stored in oeb and each stylizer was using properties from this object... meaning that each stylizer object in the code was using the last generated rules...

I updated it to store pointers to rules in the stylizer object and now, the only difference between ebooks generated before or after my changes are the generated identifiers.